adjust code in destroy action

+++ app/controllers/carts_controller.rb
@@ -17,17 +17,21 @@ class CartsController < ApplicationController
   def destroy
     cart_item = @current_cart.cart_items.find_by(product_id: params[:product_id])
 
-    if cart_item.quantity > 1
-      quantity = cart_item.quantity
+    if cart_item
+      if cart_item.quantity > 1
+        quantity = cart_item.quantity
 
-      cart_item.update!(quantity: quantity - 1)
-      @current_cart.update!(total_price: @current_cart.calculate_total_price)
+        cart_item.update!(quantity: quantity - 1)
+        @current_cart.update!(total_price: @current_cart.calculate_total_price)
 
-      render json: CartSerializer.new(@current_cart).as_json, status: :ok
-    else
-      if cart_item.destroy!
         render json: CartSerializer.new(@current_cart).as_json, status: :ok
+      else
+        if cart_item.destroy!
+          render json: CartSerializer.new(@current_cart).as_json, status: :ok
+        end
       end
+    else
+      render json: { error: 'Product must exist or was not added in the cart' }, status: :bad_request
     end
   end
